Implementing comments from @danielnelson
